# api/docs/swagger.yaml
basePath: /
definitions:
...
  proto.CreateGallery:
    properties:
      picture_url:
        type: string
    type: object
  proto.CreateTeacher:
    properties:
      bio:
        type: string
      contact:
        type: string
      experience_years:
        type: integer
      graduated_students:
        type: integer
      ielts_score:
        type: number
      name:
        type: string
      profile_picture_url:
        type: string
    type: object
  proto.GalleryRes:
    properties:
      created_at:
        type: string
      id:
        type: string
      picture_url:
        type: string
    type: object
...
  proto.GetListGalleryRes:
    properties:
      galleries:
        items:
          $ref: '#/definitions/proto.GalleryRes'
        type: array
      total_count:
        type: integer
    type: object
  proto.GetListTeacherRes:
    properties:
      teacher:
        items:
          $ref: '#/definitions/proto.TeacherRes'
        type: array
      total_count:
        type: integer
    type: object
  proto.TeacherRes:
    properties:
      bio:
        type: string
      contact:
        type: string
      created_at:
        type: string
      experience_years:
        type: integer
      graduated_students:
        type: integer
      id:
        type: string
      ielts_score:
        type: number
      name:
        type: string
      profile_picture_url:
        type: string
    type: object
info:
  contact: {}
  description: API for Pima
  title: Pima
  version: "1.0"
paths:
...
  /gallery/create:
    post:
      consumes:
      - application/json
      description: Create a new gallery entry
      parameters:
      - description: Create gallery
        in: body
        name: gallery
        required: true
        schema:
          $ref: '#/definitions/proto.CreateGallery'
      produces:
      - application/json
      responses:
        "200":
          description: Gallery created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create gallery
      tags:
      - gallery
  /gallery/delete:
    delete:
      consumes:
      - application/json
      description: Delete a gallery entry by ID
      parameters:
      - description: Gallery ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Gallery deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete gallery
      tags:
      - gallery
  /gallery/get:
    get:
      consumes:
      - application/json
      description: Get a gallery entry by ID
      parameters:
      - description: Gallery ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.GalleryRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get gallery by ID
      tags:
      - gallery
  /gallery/list:
    get:
      consumes:
      - application/json
      description: Get all galleries with optional filters
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.GetListGalleryRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get gallery list
      tags:
      - gallery
  /gallery/update:
    put:
      consumes:
      - application/json
      description: Update gallery entry
      parameters:
      - description: Gallery ID
        in: query
        name: id
        type: string
      - description: Gallery Update Details
        in: body
        name: gallery
        required: true
        schema:
          $ref: '#/definitions/proto.CreateGallery'
      produces:
      - application/json
      responses:
        "200":
          description: Gallery updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update gallery
      tags:
      - gallery
  /teachers/create:
    post:
      consumes:
      - application/json
      description: Create a teacher with details
      parameters:
      - description: Create teacher
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/proto.CreateTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: Teacher created successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a teacher
      tags:
      - teachers
  /teachers/delete:
    delete:
      consumes:
      - application/json
      description: Delete a teacher by ID
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Teacher deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a teacher
      tags:
      - teachers
  /teachers/get:
    get:
      consumes:
      - application/json
      description: Get a teacher by their ID
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.TeacherRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get teacher by ID
      tags:
      - teachers
  /teachers/list:
    get:
      consumes:
      - application/json
      description: Get a list of teachers with optional filters
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: Minimum years of experience
        in: query
        name: experience_years_min
        type: integer
      - description: Maximum years of experience
        in: query
        name: experience_years_max
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/proto.GetListTeacherRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get list of teachers
      tags:
      - teachers
  /teachers/update:
    put:
      consumes:
      - application/json
      description: Update teacher details
      parameters:
      - description: Teacher ID
        in: query
        name: id
        required: true
        type: string
      - description: Update teacher
        in: body
        name: teacher
        required: true
        schema:
          $ref: '#/definitions/proto.CreateTeacher'
      produces:
      - application/json
      responses:
        "200":
          description: Teacher updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update a teacher
      tags:
      - teachers
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"